// src/api/generated/model/sessionsFindAllSportsItem.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API v1
 * API for the Ludora app
 * OpenAPI spec version: 1.0
 */

export type SessionsFindAllSportsItem =
  (typeof SessionsFindAllSportsItem)[keyof typeof SessionsFindAllSportsItem];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionsFindAllSportsItem = {
  BASKETBALL: 'BASKETBALL',
  FOOTBALL: 'FOOTBALL',
  TENNIS: 'TENNIS',
  VOLLEYBALL: 'VOLLEYBALL',
  PADDEL: 'PADDEL',
  BADMINTON: 'BADMINTON',
  'PING-PONG': 'PING-PONG',
} as const;